+++ src/app/locales/en/components/filesystem-browser/dir-load-error.js
@@ -1,4 +1,5 @@
 export default {
   navigateToRoot: 'Navigate to root directory',
   unknownError: 'Unknown error',
+  description: 'Most often, this indicates problems with the underlying storage backend. Contact the administrators if the problem persists.',
 };