+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Latest changes
 
+* VFS-13235 Added description for EAGAIN error
 * VFS-13221 Added support for new Calendar Versioning scheme in Onedata (e.g. version 25.0)
 * VFS-13198 Added support and XML editor for OpenAIRE metadata
 * VFS-12892 Added support and XML editor for DataCite metadata